feat: Add update_photo endpoint to AssuntoController and DocumentoController
The AssuntoController and DocumentoController have been updated to include a new endpoint called update_photo. This endpoint allows users to update their photo by providing the _id of the user and the uploaded image file. The AssuntoService and AssuntoRepository have also been updated to handle the photo update logic. This change improves the functionality of the application by allowing users to easily update their profile photos.